Get correct keys and constants for dev and production environments
Reference the appropriate env variables for API keys and constants
based on NODE_ENV. If production, get the keys and constants for
the actual live services, Twitter accounts, Slack groups, etc. If
not, get keys / constants for the test accounts.

Based on setting the NODE_ENV variable to 'production' on the actual
production server.