Add number of manually invited users to constants
The Typeform we will be using to invite new members has already received 50 responses from people signing up for NYCEDU. Those folks have already been invited to the Slack team manually, and we do not want to send them invites again.

This means that the first person who fills out the form after we're deployed in production will actually be response numbe 51. So we want to send typeform an offset of 50. This would be computed as the constant number of manually invited users (50) plus the length of the `previouslyInvitedProd.json` file (0).

If more people join between now and the production deploy, the constant can be updated again.