Rename users table form_submissions
Each row in the table has an email and a date_submit, with a unique constraint only on email and date_submit together. So this is really a table of form submissions, not a table of users. (A user's email could be in multiple rows).

Refactor table name and method names throughout the db methods. I left the term `user` in methods that acutally send invitations and follow people because at that point we are talking about a specific NYCEDU user.